filterx: merge libtest/filterx and lib/filterx/tests/filterx-lib into libtest
