+++ lib/filterx/tests/test_object_integer.c
@@ -21,6 +21,8 @@
  *
  */
 #include <criterion/criterion.h>
+#include "libtest/filterx-lib.h"
+
 #include "filterx/object-string.h"
 #include "filterx/object-null.h"
 #include "filterx/object-primitive.h"
