+++ lib/filterx/tests/test_object_message.c
@@ -21,22 +21,11 @@
  *
  */
 #include <criterion/criterion.h>
+#include "libtest/filterx-lib.h"
+
 #include "filterx/object-message-value.h"
 #include "apphook.h"
-#include "filterx-lib.h"
 #include "scratch-buffers.h"
-
-static void
-assert_object_json_equals(FilterXObject *obj, const gchar *expected_json_repr)
-{
-  struct json_object *jso = NULL;
-
-  cr_assert(filterx_object_map_to_json(obj, &jso) == TRUE, "error mapping to json, expected json was: %s",
-            expected_json_repr);
-  const gchar *json_repr = json_object_to_json_string_ext(jso, JSON_C_TO_STRING_PLAIN);
-  cr_assert_str_eq(json_repr, expected_json_repr);
-  json_object_put(jso);
-}
 
 Test(filterx_message, test_filterx_object_message_marshals_to_the_stored_values)
 {
