+++ lib/filterx/tests/test_object_primitive.c
@@ -21,23 +21,12 @@
  *
  */
 #include <criterion/criterion.h>
+#include "libtest/filterx-lib.h"
+
 #include "filterx/filterx-object.h"
 #include "filterx/object-primitive.h"
 #include "filterx/object-message-value.h"
 #include "apphook.h"
-#include "filterx-lib.h"
-
-static void
-assert_object_json_equals(FilterXObject *obj, const gchar *expected_json_repr)
-{
-  struct json_object *jso = NULL;
-
-  cr_assert(filterx_object_map_to_json(obj, &jso) == TRUE, "error mapping to json, expected json was: %s",
-            expected_json_repr);
-  const gchar *json_repr = json_object_to_json_string_ext(jso, JSON_C_TO_STRING_PLAIN);
-  cr_assert_str_eq(json_repr, expected_json_repr);
-  json_object_put(jso);
-}
 
 Test(filterx_object, test_filterx_object_construction_and_free)
 {
