+++ libtest/CMakeLists.txt
@@ -11,7 +11,7 @@ set(LIBTEST_HEADERS
     stopwatch.h
     cr_template.h
     grab-logging.h
-    filterx.h
+    filterx-lib.h
 )
 
 set(LIBTEST_SOURCES
@@ -28,7 +28,7 @@ set(LIBTEST_SOURCES
     stopwatch.c
     cr_template.c
     grab-logging.c
-    filterx.c
+    filterx-lib.c
 )
 
 add_library(libtest STATIC ${LIBTEST_SOURCES})
